pipelines-batch/expansion_strategy/trusted_zone/requirements.txt
pyspark==3.4.1.
pymongo
python-dotenv
pandas
hdfs
